almost done
The things I've done are:
add remove function for the many todays.
Able to manitpulate dom and show up appropriate info with input.

Things I have to do:
There's a bug. When I click the random button. which adds the many days
it removes the text content of city name and main weather.
also needs to link the submit location with the api linking(insteaed
of the button)
Finally just do some css and remove placeholder colors